Revert "epoll geht los"
This reverts commit e9ae7f5b715a6d53900c2f0b59bcb2c915914629.
