+++ example/pom.xml
@@ -148,13 +148,13 @@
       <groupId>io.netty</groupId>
       <artifactId>netty-transport-native-epoll</artifactId>
       <version>${project.version}</version>
-      <classifier>linux-${os.detected.arch}</classifier>
+      <classifier>${jni.classifier}</classifier>
     </dependency>
     <dependency>
       <groupId>io.netty</groupId>
       <artifactId>netty-transport-native-kqueue</artifactId>
       <version>${project.version}</version>
-      <classifier>osx-${os.detected.arch}</classifier>
+      <classifier>${jni.classifier}</classifier>
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
