--- example/src/main/java/io/netty/channel/epoll/EpollTunChannelConfig.java
@@ -1,9 +1,0 @@
-package io.netty.channel.epoll;
-
-import io.netty.channel.FixedRecvByteBufAllocator;
-
-public class EpollTunChannelConfig extends EpollChannelConfig {
-    EpollTunChannelConfig(final AbstractEpollChannel channel) {
-        super(channel, new FixedRecvByteBufAllocator(2048));
-    }
-}
