+++ example/src/main/java/io/netty/example/tun/TunEchoDevice.java
@@ -5,11 +5,9 @@ import io.netty.channel.Channel;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.EventLoopGroup;
-import io.netty.channel.epoll.EpollEventLoopGroup;
-import io.netty.channel.epoll.EpollTunChannel;
 import io.netty.channel.kqueue.KQueueEventLoopGroup;
 import io.netty.channel.kqueue.KQueueTunChannel;
-import io.netty.channel.socket.TunAddress;
+import io.netty.channel.kqueue.TunAddress;
 import io.netty.util.internal.PlatformDependent;
 import io.netty.util.internal.StringUtil;
 
@@ -45,19 +43,20 @@ public class TunEchoDevice {
     public static void main(String[] args) throws Exception {
         EventLoopGroup group;
         Class<? extends Channel> channelClass;
+        if (PlatformDependent.isLinux()) {
+            group = new EpollEventLoopGroup(1);
+            channelClass = EpollTunChannel.class;
+        }
         if (PlatformDependent.isOsx()) {
             group = new KQueueEventLoopGroup(1);
             channelClass = KQueueTunChannel.class;
-        }
-        else if (!PlatformDependent.isWindows()) {
-            group = new EpollEventLoopGroup(1);
-            channelClass = EpollTunChannel.class;
         }
         else {
             throw new RuntimeException("Unsupported platform: This example only work on Linux or macOS");
         }
 
         try {
+            java.lang.Class<? extends Channel> channelClass = KQueueTunChannel.class;
             final Bootstrap b = new Bootstrap()
                     .group(group)
                     .channel(channelClass)
@@ -85,9 +84,6 @@ public class TunEchoDevice {
                 else {
                     throw new RuntimeException("Unhandled address type: " + ADDRESS);
                 }
-            }
-            else if (!PlatformDependent.isWindows()) {
-                throw new RuntimeException("Implement me");
             }
 
             System.out.println("Address assigned: " + ADDRESS.getHostAddress());
