--- transport-classes-epoll/src/main/java/io/netty/channel/epoll/AbstractEpollMessageChannel.java
@@ -1,7 +1,0 @@
-package io.netty.channel.epoll;
-
-abstract class AbstractEpollMessageChannel extends AbstractEpollChannel {
-    AbstractEpollMessageChannel(final LinuxSocket fd) {
-        super(fd);
-    }
-}
