+++ transport-classes-kqueue/src/main/java/io/netty/channel/kqueue/TunAddress.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.channel.socket;
+package io.netty.channel.kqueue;
 
 import java.net.SocketAddress;
 import java.util.Objects;
