--- transport/src/main/java/io/netty/channel/socket/TunChannel.java
@@ -1,6 +1,0 @@
-package io.netty.channel.socket;
-
-import io.netty.channel.Channel;
-
-public interface TunChannel extends Channel {
-}
